Fix {i,o}stream support for pcg128_t
We defined operator{<<,>>} for pcg128_t in pcg_extras, assuming Koenig
lookup would find these as needed.  This isn't actually true, because
pcg128_t is a typedef, and Koenig lookup doesn't apply to the
namespaces of typedefs.  We would need to define those operators in
the global namespace; that's a bad idea.

Instead, work around it.  Define explicit Input and Output functions
in the namespace, which in turn default to operator<< and operator>>
if they need to, and explicitly use them in engine printing/parsing
functions.