Remove unneeded headers to keep iostream from pulling in 10^5  bytes
Admittedly it's a bit brute force to just remove the header inclusions, and that it works is based on all the iostream -related code being in templates, so that when they're not used there's no problem. Of you do use them, you have to include <iostream> yourself (and I'm not 100% sure that there won't be restrictions in order of includes).

Another approach, maybe more robust but less elegant, would be to #ifdef out the iostream headers if some flag (PCG_DONOTUSEIOSTREAM ? ) is defined

Removing <new> is not absolutely necessary, but I'm not sure what it's used for here, the only relevant part I could find is one use of placement new?