make PCG portable across compilers for non-compliant seeders
- According to the GCC standard (§8.3.6): "The order of evaluation of function arguments is unspeciﬁed."
This caused problem when PCG was initialized from another generator. GNU GCC evaluated from right (v9.3), Apple Clang evaluated from the left (v12).
The PCG was thus non-portable for non-compliant seeders (not returning the same values) as with the same seed it produced different results on various platforms.

More on this:
- https://en.cppreference.com/w/c/language/eval_order