Improve emulated 128-bit math performance (MSVC)
PCG's emulated 128-bit math was designed to provide 128-bit math
support for 32-bit systems, and as such prefers to use 32-bit types
most of the time, as these types are directly supported by the CPU
(and in general, when providing k-bit math prefers, it prefers
k/4-bit types).

The original idea was that this code wouldn't be necessary on 64-bit
systems, as 64-bit systems should provide 128-bit integer types (much
as 32-bit systems provided 64-bit types).  This assumption was true
for compilers like gcc, clang, and icc.  MSVC, on the other hand,
does not have 128-bit integers in 64-bit mode and so falls back to
using this code, but, as a result, performed no better (and often
slightly worse) than 32-bit code.

This change fixes the issue by adding 64-bit specializations,
allowing performance of 128-bit math to approach that of native
support for these types in GCC and Clang.  As such, it offers a
more than 10x speed boost to pcg64 when compiled by MSVC in 64-bit
mode.  This massive improvement is possible because the previous
performance was diabolically bad.